Simplified. Moved to views instead of owned for manipulating arrays
